requirements.txt
numpy==1.19.0
pygame==1.9.6
PyOpenGL==3.1.5
